+++ chen_cppcguilib/basicImage.h
@@ -6,14 +6,14 @@ class basicImage : public component
 {
 public:
     basicImage(const std::vector<cgui::string>& image);
-    int getWidth() const override;
-    int getHeight() const override;
+    size_t getWidth() const override;
+    size_t getHeight() const override;
     std::vector<cgui::string> getData() const override;
     
     void setImage(const std::vector<cgui::string>& image);
 
 private:
     std::vector<cgui::string> image;
-    int width;
+    size_t width;
     void calculateWidth();
 };
