+++ chen_cppcguilib/multiLineText.h
@@ -6,8 +6,8 @@ class multiLineText : public component
 {
 public:
     multiLineText(std::vector<cgui::string> textByLine);
-    int getWidth() const override;
-    int getHeight() const override;
+    size_t getWidth() const override;
+    size_t getHeight() const override;
     std::vector<cgui::string> getData() const override;
     void setMultiText(std::vector<cgui::string> textByLine);
